+++ sanic/__init__.py
@@ -1,7 +1,12 @@
-from sanic.app import Sanic
-from sanic.blueprints import Blueprint
-
-
 __version__ = "18.12.0"
 
 __all__ = ["Sanic", "Blueprint"]
+
+
+def __getattr__(name):
+    if name == "Sanic":
+        from sanic.app import Sanic
+        return Sanic
+    if name == "Blueprint":
+        from sanic.blueprints import Blueprint
+        return Blueprint
